Refactor the code. Add Bundlex.Precompiled module
